tests/wpt/css-tests/css-writing-modes-3_dev/html/writing-mode-parsing-001.htm

<!DOCTYPE html>
<html><head><title>CSS Writing Modes: parsing writing-mode</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii">
<link href="https://drafts.csswg.org/css-writing-modes-3/#block-flow" rel="help">
<meta content="This test asserts the parser and getComputedStyle works correctly for the writing-mode property." name="assert">
<meta content="dom" name="flags">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

</head><body><div data-expected="horizontal-tb" title="writing-mode initial value"></div>

<div data-expected="vertical-rl" style="writing-mode: vertical-rl"></div>
<div data-expected="vertical-lr" style="writing-mode: vertical-lr"></div>

<div style="writing-mode: vertical-rl">
  <div data-expected="vertical-rl" title="writing-mode should inherit"></div>

  <div data-expected="horizontal-tb" style="writing-mode: horizontal-tb"></div>

  <div data-expected="horizontal-tb" style="writing-mode: initial"></div>
  <div data-expected="vertical-rl" style="writing-mode: inherit"></div>
  <div data-expected="vertical-rl" style="writing-mode: horizontal-tb; writing-mode: inherit"></div>

  <div data-expected="vertical-rl" style="writing-mode: foo"></div>
</div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).writingMode;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
</body></html>
